Metadata streaming dialog: use a new flag, utilize metadata connector, new metadata list. re #44
Because Studio API wrapper is undergoing major redesign, do not pass this function object anymore when opening metadata streaming dialog from places other than add-on settings. A new flag will replace this object instead.
In splmisc module, add a new metadataList function that'll format metadata streaming values as a list for use in various modules and functions.
In metadata dialog, instead of appending to metadata list, build the streams list via comprehension. The list will be generated early so it can be assigned directly to metadata list in the add-on settings dialog and/or to be passed in as list of servers in splmisc.metadataConnector function and applied to current broadcast profile if told to do so.