Studio API: let the API wrapper return time values which will then be passed onto time nanouncer. re #45.
Part of API wrapper redesign: one of the arguments for StudioAPI function is a function object. So far, the only time it was used was for time announcements, which makes it a dependent argument. From now on, the API wrapper will return time value, which will then be fed to time announcer, makng the wrapper more independent. Thus, passing a function object in API wrapper is deprecated and will be removed in a future commit.