Time range finder/Studio API wrapper (18.02): call Studio API from splbase module when looking for tracks. Re #45.
Instead of passing in StudioAPI function object, let time range finder work directly with the wrapper in splbase module. This makes the code look cleaner and the dialog less dependent on the function passed in. At the same time, use select track function from splbase when a track is found.