15.13-LTS: refminements to metadata announcer during profile switches.
In old days, when metadata status was announced during profile switches, NVDA would sleep for two seconds. This meant commands would not respond and NVDA may not appear to do anything. Changed it so wx.CallLater will deal with this, along with an internal function (nside metadata announcer) to announce metadata status if asked.