Add-on update: a framework to deal with update errors is now available.
A set of constants to deal with error conditions:
* 0: no errors.
* 1: generic error (reserved).
* 2: NVDA is in secure mode.
* 3: this is a custom try build.
* 4: NVDA is run from source.
* 5: This is a Windows Store NVDA (no add-on updating supported).
* 6: NVDA has an add-on update feature built in.
* 7: no internet connection.
Of these, 2 through 6 are major errors, so no channel selection, retrys and others are not supported under such conditions. These constants are subject to change.
To make these error messages friendly, a dictionary to store error constants and user friendly messages is used to convey error information.
A new splupdate.canUpdate function has been added, which is a convenience method to check splupdate.IsAddonUpdatingSupported function and checking if the result is 0 (no errors). For the most part, this convenience function will not be used except for background update checking.