Update check (18.02/15.13-LTS): provide a more generic error message if things other than known errors occur.
If a generic error occurs, the best thing to do is telling users to look up the log viewer. Thankfully, a message already exists to deal with this one, so use this instead of the custom message used until now.
Also, when generic error occurs, print the resulting traceback no matter what log level one is using (log.error not log.debugWarning).
This will be backported to 15.x LTS.