Profile flags: profile switch start/end methods, profile flags dictionary.
A new private dictionary will be used to store switch profile flags (0x1 for instant switch, 0x2 for timed switch for now).
When it comes to profile switching, until now there was no start/end methods to set flags, cache profiles and so on. Starting from add-on 18.04, the new switch profile start/end methods will be used to enter or exit switch profiles. These will check profile switch flags to make sure one isn't tampering with flags and to set appropriate flags for the correct switch type. Also, flag checker is now a property, which allows more flag checks such as making sure the prev profile is properly set.
As for profile flags, it'll be a bit array, with xor operations used to toggle flags on or off.