Update check: don't check for uypdates if profile switch flags are on, register/unregister action handler.
Because update module will be initialized after the config database, take advantage of this by registering the action handler and making sure trigger flags are turned off (if on, update chekcing will not happen).
Also, the action handler for update facility is hereby defined - just an empty code for now, to be spelled out in the next few commits.