Profile switching/metadata status: do not allow status to be announced multile times. Re #51.
Consider the following scenario: if an instant switch profile is defined and the user switches back and forth between current profile and an instant switch profile. If this happens, one will hear metadata status announcements multiple times, ultimately caused by not chekcing if the timer is running or not. To work around this, threading.Timer object will be used to keep an eye on this. Whenever profile switch is made, the former timer object will be killed, to be replaced by a fresh new timer. This prevents multiple status announcements.
Because it affects both stable and LTS, this will be made avaialble for 15.x users. And due to internal differences, this commit will be for 15.x, while a more elegant one is planned for a future stable release.