Profile triggers (18.03/15.14-LTS): also consider profile switch duration. re #52.
Previously, the app module did not care if the duration was 0 or something else. Now it will: if duration is specified for a time-based profile, the show profile will take over as long as it is still happening.
Also, instead of returning a 3-tuple, next timed profile function will add a fourth element: duration delta in seconds, used to calculate profile switch timer interval correctly. Without this, the end timer will use the original duration, which leads to profile switching failing whenthe show is over.
This will be backported to 15.x-LTS.