Profile triggers (18.03/15.14-LTS): restart trigger end timer manually if trigger subsystem is told to restart.
If triggers subsystem is told to restart, it wouldn't restart the active timer, thereby leading to issues such as switching back to regular profile at wrong times, not responding to updated trigger settings and so on. From now on, whenever tre it is told to restart and if it happens that the trigger end timer is running, it'll be restarted.
This commit also takes care of a situation where profile triggers wouldn't start after closing add-on settings once update trigger settings were registered. There is still an issue with this not working when Cancel button is pressed form add-on settings, and that is next.