Profile triggers: let switch profile function run so profile switching can occur when Cancel button is pressed from add-on settings.
If add-on settings is open and it is time to broadcast a show, when Cancen button is clicked, profile switching didn't occur. This is due to the config database's switchProfile function checking for dialog open flag in UI module, but the Cancel method in the add-on settings does not clear this flag to the end. Now fixed so the flag will be cleared as soon as onCancel event handler is entered so profile switching can occur.