Metadata streaming: metadata list function no longer takes Studio handle, raise error if handle itself isn't found.
Because theo nly use of handle in metadataList was for convenience, this function will no longer take any arguments. Also, instead of using SendMessage function, obtain results via splbase.studioAPI function. The benefit of this move is that the list will bne filled with NULL if Studio Handle itself is NULL, which then allows the function to raise a runtime exception which should be handled by callers.