Metadata streaming: allow metadata announcer to resume action handler if Studio handle isn't ready yet a time-based profile is active.
As a follow-up from the previous commit: if Studio handle is not ready, set a flag in splmisc module to tell metadata announcer to resume action handler. This will be the case if a time-based profile is up when Studio handle isn't and NVDA is told to announce metadata status whenever profile switch occurs.