Partial playlist snapshots (18.05): introduce ability to take playlist snapshots in the middle of a playlist. Re #55.
Playlist snapshots is useful for viewing info about a complete playlist. But what if a broadcaster wants to view info for a specific hour, or from the current location to the end of the playlist? This is done through partial playlist snapshots in combination with playlist analysis (renamed from track time analysis). Thus, a broadcaster would mark a track for start of analysis, move around the playlist, then take a snapshot from the current location to the marked track, similar in concept to old track time analysis (now a part of a suite of features on playlist statistics and analysis, which will include other features).
Due to extensive testing required, this will be tagged for 18.05.