Playlist transcripts (18.05): fetch column data for all columns at once, necessary foundaiton for playlist transcripts. Re #56.
As part of playlist transcripts, trakc item object iwll gain a new method to return column data for all columns at once. This is mostly meant for debuggoing purposes and planned to be used by playlist transcripts, a way to obtain playlist info and export it to a variety for formats (including clipboard).