Config UI: add foundaiton for apply button handler. Re #59.
NVDA Core will now include 'apply' button to apply current settings to the active profile immediately without having to reopen the preferences dialog in question. SPL add-on will follow a similar path where add-on settings will apply to selected profile (with a warning if needed). For now, apply button is a copy of OK button handler.