Columns Explorer (18.06): manually build columns list when closing the dialog in order to avoid a reference problem. Re #62.
Discovered while working on multi-category settings: suppose a broadcaster changes column positions in Columns Explorer dialog in add-on settings, and then clicks Cancel form main add-on settings window. The position should not have changed, but it did. This is caused by references - a reference to a list is still a reference to a list, and it'll be a reference until a new list is created or the data type changes.
To get around this, manually build column positions list when OK button is pressed from Columns Explorer dialog.