Columns Explorer: now a category in add-on settings, applied the fix from column position reference issue. Re #6.
Columns Explorer is now a settings category in add-on settings. For now, previous behavior (a separate config dialog) will be kept, but eventually the combo boxes will be shown on the main add-on settings screen and will be grouped under 'Studio' and 'Track Tool' groupings.
Also, the fix from issue 62 (list reference problem causing settings to be retained even with Cancel button activated) is now available.