Playlist transcripts: allow playlsit contents to be copied to the clipboard in text format. Re #56.
Another milestone: playlist content can be copied to the clipboard in text format. To support this, a new combo box is introduced in Playlist Transcripts dialog to select transcript action. This also means introduction of transcriptAction argument in transcribers so it can perform actions based on action flag.
Not all transcript formats support copying data to the clipboard. For now, text format supports this.
The action combo box also includes save to file option which will be added in the future.