Playlist transcripts: let users save text transcripts of the playlist to a file. Re #56.
Milestone 3: if the broadcaster chooses to save plain text playlist transcripts into a file, the file will now be created under the user's Documents folder under nvdasplPlaylistTranscripts folder. The filename is yyyymmdd-hhmmss-playlistTranscript.txt. This whole thing is performed by a new function created for this purpose that'll save transcripts to a variety of formats based on extension parameter.