Track item reorganization: copy column number variable from Studio's track item class. Re #65.
In preparation for moving some scripts to the abstract SPLTrackItem class, copy column number variable to Track Tool and SPL Creator classes. This also saves a lookup round as all that needs to be done when moving through columns is consult this column number variable, not the one hosted by the app module instance.