Track item class reorganization: move Columns Explorer to the base class. Re #65.
Because all track items for various SPL suite of apps are the same (excpet for minor differences), it meant COlumns Explorer was duplicated across classes 9particularly between Track Tool and Creator, which are identical). Thus move this functionality to the base SPLTrackItem class. This also means overlay class initializer is now part of the base class as well.