Track items and tracks list: now powered by SysListView32.ListItem and SysListView.List, respectively. Re #69.
The biggest impact of using SysListView32 objects is availability of table navigation commands, and in case of the list itself, column count and order array. This means track items will be recognized as table rows. Becauase of backward compatibility, some aspects of this will change to suit the needs of individual item classes, namely playing table edge tone instead of providing speech and braille feedback, vertical column navigation setting and what not.