Broadcast profiles panel: add a flag that records currently selected profile.
In preparation for implementing onPanelActivate method in profile sensitive panels (metadata streaming, for example), a flag will be used to record the name of the selected profile. This flag will be NULL if the selected profile is the active profile, otherwise it'll record the name of the selected profile. In case the profile is renamed and if it happens to be the selected profile, the new name will be reflected on the flag. If the profile is gone, the flag will be nullified.
Also added translator comments to broadcast profile panel title text.