Profile specific panels: save settings to appropriate profile. Re #6.
When saving/applying profile settings, consult selected profile flag to determine which profile should be modified. This is way simpler than the old interface's aplicability check, as in that saving settings can be done in one pass rather than going through a conditional. Due to risks involved such as possible regressions, this optimization won't make it to old interface.