Playlist transcripts settings: extend column announcements dialog to handle column selector settings for Playlist Transcripts. Re #56.
Requested by a broadcaster: add settings for plyalist transcripts, including which columns to include and what not. This is achieved by extending column announcements dialog to include a flag that change dialog text and other behavior so playlist transcripts/column selector settings can be changed via this dialog. This change is part of old add-on settings interface."
"