Config action: listen to config.postConfigSave action. Re #73.
Until now in order to save add-on settings, one had to close the app module (or call save volatile private function). This means save command (Control+NVDA+C) could not be used. Thanks to config.postConfigSave action, one can now save add-on settings from everywhere by pressing Control+NVDA+C or choosing 'save configuration' from NVDA menu.
The handler fopr this action will call the private save volatile function with a flag that tells ConfigHub to not set volatile config flag to True, which then allows config to be saved in the future when Control+NVDA+C is pressed in the future. The volatile function was chosen because config action is meant to be invoked while STudio is running, which means not closing profile switching history, not clearing new profiles set and what not - in other words, keeping ConfigHub alive. Thus, ConfigHub.save will be reserved for app module termination function, which then will clean up in-memory flags such as new profiles set and what not.
This is destined for a version to be released once NVDA release with postConfigSave action included.