Broadcast profiles: an explanatory text will be shown if only normal profile can be used. Re #56.
In case normal profile is in use (set via command-line or if in-memory copy is specified), present an explanatory text. Although the label for it is empty (as far as how NVDA sees it), the value text will tell users that only normal profile is in use. This is the compromise approach after discovering that category classes cannot be inserted in real-time.