Deprecated keys: store them in a tuple.
Until now when it comes to deprecating keys, a static dictionary was used. This had one problem: multiple keys from the same section could not be removed. To get around this, deprecated keys will be stored in a tuple, and these will be represented in the string of the form section/key. This then allows multiple keys from a single section to be removed.