Add-on manifest: keep a reference to manifest dictionary.
Keep an eye on add-on manifest. This is needed so add-on update channel can be changed through the manifest (not advised to do so due to special nature of manifest files, but this is one of the ways to make the add-on conform to future changes in regards to add-on updates and what not).