Add-on settings (18.09-LTS): retain temporary setting changes so they can show up when switching between profiles and/or panels. Re #77.
Consider the following scenario: a user opens add-on settings, makes changes to profile specific settings, and switchges panels. When the user returns to the just changed setting, the value is lost.
This is caused by not remembering setting changes for broadcast profiles. Thus introduce a hidden temp map that'll record current settings (changes) and reloads it when the panel becomes activated. Depending on whether the map contians the name of the active profile, it'll load temp settings (True) or use profile setting (False).
For now, this change has been applied to metadata streaming list.