Pilot features: restore pilot features checkbox, pilot features flag.
Pilot features checkbox was removed as a result of handing over update facility to Add-on Updater (NVDA Core issue 3208), thus restore this.
Also, a new flag has been added to Config Hub to check if pilot features can be used, as well as a flag to actually enable Test Drive/pilot features.