Update check/default flags: use manifest routines to set default update channel/URL variables.
In order to allow releases to be made from any branch (removing release branch along the way), use add-on manifest's updateChannel and version metadata to set default update channel and URL flags in update module to dev or stable, respectivley. Also, create a boolean flag to indicate if dev version is in use (applicable if a true development version iis in use, although this may require tweaking in case custom try build is in use).