Update check: no more 'pending channel change' in app module.
In the past, when update channel selection was included, Studio app omdule had to check this flag in order to not proceed with update checks (it may get stuck without checking this flag, see issue #9). Now that channel selection the coresponding channel change flag is gone, just check for updates if allowed.