Encoder connection: catch wxPython assertion error in order to prevent text output from failng, causing connection routines to fail. Fixes #79.
wxPython 4 change: wxPython Phoenix is more strict about where timers can be invoked from (main thread only). Because encoder connection routines are done from another thread, this runs the risk of causing assertion error. This was observed on some speech synthesizers and/or braille displays, and to get around this, enclose the entire encoder status message method inside a try block.
This is destined for LTS18 as well.