Add-on settings: do not clear current settings map when Apply button is pressed. Fixes #80.
Oversight: when Apply button is pressed from add-on settings screen, current settings map form profile specific panels is nullified. This causes subsequent presses of OK and Cancel buttons to cause an error, thus not allowing the screen to close. Thus detect if Apply button is pressed (via the global flag) and do not clear settings map accordingly.
This fix is also destined for LTS18.