Config save action: register and unregister config save action unconditionally.
Config save action is now a requirement, making this add-on compatible with NVDA 2018.3 and later.
Also, add a call to unregister function for config save action to signal that config save routine won't be called again (unless a really obscure bug makes this assumption false).
This change is destined for LTS18 in 2019.