Metadata streaming (18.11): NVDA will no longer appear to do nothing or play error tones when trying to announce metadata stremaing connection message when Studio starts.
Regression: scope resolution problem: metadataStatus is part of splmisc module, so make sure to include that namespace. This fixes an issue where if metadata streaming announcement is set to be announced when Studio starts, an error tone or nothing will be done instead.
This patch will be backported to LTS18 as that too is affected by this regression.