Metadata status announcement: announce it at startup if NVDA restarts while Studio is running and focused in places other than playlist viewer. Re #83.
Consider the following: user runs Studio, opens Options or other screens, then restarts NVDA. If NVDA is told to announce metadata streaming status at startup, with this condition active, NVDA won't announce this because it is waiting for foreground event on playlist viewer to be activated.
Thus see if NVDA has restarted while STudio is running, and if so, just announce current metadata status. This is done by asking NVDA for a foreground object, and if this is queried while NVDA is starting, it'll be None (not defined), as NVDA's notion of foreground is a global flag.
This will be backported to LTS18 and no readme entry is necessary.