Place marker track/playlist dispatch (18.12): present generic error message.
Syntax error: possibly due to an unrecognized character in the custom message, perhaps caused by copy and paste operations or pressing a wrong key. Fixing this requires removing that message, and it also gave na opportunity to think about error message for place marker. Thus in the end, the generic error will be announced through the central playlist dispatch.