Carts Without Borders: simplify formula, use Studio 5.30 formula if needed. Re #85.
Studio 5.20 and earlier uses data words for cart key construction, whereas Studio 5.30 and later can use cart bank index. In summary, the formulas are:
* Old: (64K * cartIndex from 1 through 24) + (modifier from the order None, Shift, Ctrl, Alt) + 1
* New: (modifier * 24) + cartIndex
For example, F5 cart key is constructed thus:
* Old: 64K * 5 + 0 (None) + 1 = 327681.
* New: (0 (None) * 24) + 5 = 5.
The new formula is easier to understand, but Studio 5.30 also supports the old way too. Thus detect which formula to use based on Studio version.
Also, while adding support for the new formula, refactor old formula to add modifier value instead of doing a bitwise or (|) to make it easier to follow the algorithm.
This work is also destined for LTS18.