Playlist analysis check/dispatch (18.12): more generic message, now powered by playlist dispatch. re #81.
The checking routine for playlist analyzer will use results from central playlist dispatch so it can provide more meaningful error messages. Spekaing of error messages, it has been more generalized, as playlist transcripts is now under the playlsit analyzer umbrella.
Also, a new boolean flag is available to check for presence of last focused track flag. This is absolutely needed by track time analysis, as it needs to know where to begin and where to stop. In case partial playlist snapshots is needed, it must know the snapshot range, and in the future, partial transcripts will require this flag.
Due to extensive changes, this is destined for 18.12 and later.