StudioIsRunning: add additoinal handle presence checks.
Previously, splbase.studioIsRunning function returned the status of private SPL window handle variable. However, there are more cases where Studio may say it is alive or dead.

Specifically, the following conditions will be checked:
* Studio window handle is defined and user32.dll::FindWindow returns a valid handle for Studio.
or
* Studio handle is not defined yet user32.dll::FindWindow says a valid handle is present.
The layout of the code is such that the private window handle varialbe will be checked first, and FindWindow function will be callced if the specified condition is true.
This change is exclusive to 19.02 and later.