wx.PyTimer: NVDA will no longer papear to do nothing or play error tones if microphone is turned on an mic interval is set.
Yet another wxPython 4 compatibility problem: again wx.PyTiemr must be started from the main thread. In case of microphone interval timer, this meant nVDA would papear to do nothing or play error tones if interval is set (the timer to be run after the initial mic timer is run and used to remind broadcasters that microphone is on).
This fix (along with changes made to remaining wx.pyTimer objects) will be backported to LTS18.