Config Hub: employ configobj.get_extra_values function to remove deprecated keys in one go. Re #95.
Configobj 4.7.0 and later ships with get_extra_values function that'll return a list of 2-tuples denoting extra section(s) and key(s). Prior to this (in NVDA releases prior to 2018.4), one had to keep a static list of deprecated keys and remove each entry if it exists in add-on settings (each entry was stored in section/key format). Now that the new function is employed, there is no need to keep the static list around, and the overall key removal loop has been simplified by removing only the keys that are actually present.
There are certain keys/sections that are needed for various operations. Thankfully, they are part of the root section (section tule is empty), so they can be skipped (unless a future setting is part of an existing section). For one or two of them, do not delete it (instant switch profile flag, for instance).
This will be backported to LTS18.