Config reset: a confirmation message will be presented if instant switch and/or time-based profile is active. Re #96.
Stemming from config reset work (19.03): if an instant switch and/or time-based profile is active and settings are reset to defaults, switch flags will be left in unpredictable state. Thus make sure the user won't reset settings if this is the case, and if the user do say settings should be reset, reset switch flags, turn off timers and so on.
This will be backported to LTS18.