ConfiUI: add a base panel for column announcements handler. Re #97.
A new column announcements base panel will be used just for handling column order changes. This abstract base panel (to be declared as abstract in the future) is responsible for dealing with column announcement setting changes, most notably column order up/down button handling. As this is an abstract panel, it cannot be instantiated - any panel that deals with rearrange control can subclass this panel (until a more usable way to work with wx.RearrangeCtrl is found).
This change is exclusive to stable channel in 2019.