Config dialog: Advanced options -> Advanced.
NVDA Core now adds 'Advanced' settings panel. To align with this change, Studio add-ons settings's own advanced options panel will be renamed to 'Advanced'. For now the add-on will define its own translation for this, but in a future release that requires NVDA 2019.1, it'll be pulled straight from NVDA.
This change will be backported to LTS18 with localization edits.