SPL Utils: do not load the global plugin if secure screens and AppX is in effect.
Security and compliance: if one is running NVDA in secure mode and/or if Windows Store version is in effect, global plugin should not work. This change also allows scripts (with docstrings) to assume that NVDA is in normal mode (removing secure mode checks from each script).
This will be backported to LTS18.