SPl Utils: simplify conditional for focus to Studio script.
Logic change in focus to Studio script: previously, there were two conditional branches:
1. Don't do anything if focused on Studio window.
2. Do switch to Studio window if not focused already.
This was spelled out via if/else. Becuse the first branch was just a return statemenet, simplify it by unconditoinally moving onto the second branch if not focused on Studio window. Although the code looks different, it still presents the same logic.
This subtle change is also destined for LTS18 if needed.