SPL Config: do not close ConfigHub if another instance of the same SPL component is in use. Re #99.
Suppose a user runs Track Tool or SPL Studio and runs another instance of this. When the second instance closes, ConfigHub is also closed. Thus detect if other instances of the given SPL component is running before deciding to remove the component from the SPL components registry.
This will be backported to LTS18.