SPL app module: create an abstract indexOf functoin in SPLTrackItem class. Re #103.
In order for a class to become abstract, one or more methods in it must be marked with abstractmethod decorator. In case of SPLTrackItem class, create an abstract indexOf method that will 'return' None i.e. causes columns to not be announced, forcing concrete implementations to spell out this method.