Creator and Track Tool: obtain actual column number via track item index of method directly. Re #106.
Because Creator and Track Tool's column anouncer uses the same idea as Studio track item for fetching headers, edit the source code so they now resemble each other. Also, as an optimizatoin bonus, fetch the column header/number directly via indexOf method instead of going through a conditoinal, which duplicates the work done by the indexOf method anyway.
As this is a foundaiton for further column announcer work, this is destined for 19.07 stable only.