Config dialog: send profile switch notification whenever OK button is pressed or settings are applied to the active profile. Fixes #108.
Consider the following: a broadcaster opens add-on settings while microphone is on, changes mic alarm settings, and presses Apply or OK. If this happens with microphone alarm on, the alarm won't stop. This, along with potential metadata streamer server connection problem, is caused by the relevant subsystems not receiving profile switch notification, which is used to tell them to apply changed settings and act accordingly. Now this action will be notified via postSave method in broadcast profiles panel.
This will be backported to LTS18.