SPL Studio track items: standardize around indexOf function. Fixes #109.
Because columns can be reordered on screen but not in memory, a private original index function was provided to allow each SPL Studio track item subclass to provide the original index representation of columns. Now taht things have standardized around SPL 5.10 items, and because the only time original index is claled is index of function in track item class, combine the private and public-facing functions together, thereby standardizing around index of function (already done for Creator and Track Tool but not in Studio proper).
This is destined for 19.08 and later.