Studio app module: do not proceed with re-constructing the app module if wx.App fails. Fixes #110.
Consider the following: user opens Studio add-on settings, and then restarts NVDA without closing add-on settings. This may cause assertions to be logged, related to wx.App. This is caused by an attempt to queue Studio window handle locator thread from the main thread when the app is gone. Thus catch this and do not proceed with app modulie reconstruction when wx.App returns NULL.