Broadcast profiles: restart trigger timer. Fixes #111.
During migration from old dialogs collection to settings panels, trigger timer restart routine was somehow forgotten, and now it is back. just with Cancel button handler, trigger timer will be restarted when OK button is clicked i.e. when the screen is being closed.
This will be backported to LTS18.