SAM Encoders: terminate connection reporting thread if status object is gone.
If background monitoring is on and the encoder entry in SAM Encoders is deleted, Python thros type error because message cache checker looks for a string when it can't find it. This is caused by not checking if encoder status object's name is NULL or not. Thus terminate the reporter thread if status object's name (encoder status) is truly gone.