Encoders support: transfer to SPL Engine app module package. Re #98.
Since late 2014 SPL add-on included encoders support module as part of SPL Utilities global plugin. Back then it was a global plugin due to anticipation that only a few apps that uses these classes are to be found. Now that all of them were found (SPL Engine and Streamer), there is no need to keep it as a global plugin - rather, it is now part of base SPL Engine app module package, and by extension, Streamer app module as well. With this transfer, the scope of SPL Utils has been narrowed a lot - only providing Studio and Engine support services, mostly SPL Controller layer.