Announce column content: use visual column order/layout if requested. re #117.
As part of generalizing Columns Explorer, a new keyword argument (visualColumns) is introduced in SPLTrackItem.announceCOlumnContent method. If set to yes (default), the column the method receives will be treated as true column number as opposed to header index. This allows Columns Explorer to work properly in more more places.