Copy profile: work directly with parent's profile names and selection when displaying and selecting base profile for copying, respectivley.
When displaying base profiles for profile copying, an iterative string split was performed via list comprehension. Although this operation is fast, it overlooks profile names list from add-on settings/broadcast profiles category, which actually holds profile names ready for use for this kind of situation. In addition, selection index went through another string split, this time based on the name of the profile selected from the parent screen, adding more bytecode instructions.
Thus the following changes:
* parent.profileNames will be consulted directly when displaying base profiles. This avoids unnecessary iterative string splits.
* When selecting which base profile to use, index from the selected profile in add-on settings will be used directly, avoiding a string split operation.
Combined, this results in less bytecode instructions and makes the code look cleaner. Although it might be possible to profile names list and selection index might be tampered with, for now changes in this commit should suffice (besides, when new/copy profile dialog is invoked, add-on settings (parent) is disabled).