Metadata streaming dialog: pull in metadata streams from connections list.
Dead code elimination: until 2018, metadata streaming dialog was used for two purposes: to set metadata streaming options globally, and to change connection servers on the fly without opening add-on settings (and applying changes to add-on configuration if told to do so). Since 2018, the first aspect is done through add-on settings under metadata streaming category, so only the second aspect remains. Threfore:
* Remove dialog text associated with add-on settings.
* Pull in connection server details from connected servers list (splmisc module).