Metadata streaming dialog: change streams and optionally apply it to add-on settings database.
Dead code elmination: in OK button event handler, change connection settings for metadata streams, and optionally apply changes to add-on settings if told to do so. This eliminates calls to dialog.Parent for saving metadata stream settings to add-on settings, as that aspect of metadata streaming dialog is gone. Also this change is applicable to Cancel button event handler as well.