Encoders: define 'announce connection status until connected/connection stop on error' flag. Re #122.
A new flag will be used to contorl whether NVDA will stop announcing connection status when an error is first encountered. In the flag map, it'll be known as 'connection stop on error', but in the encoder properties it'll be seen as 'announce connection status until connected'. This is a deliberate decision: the flag map name is used to store encoders for which NVDA should give up after an error is encountered, and is intended to record archive encoders. But when it comes to actual connection reporter routine, the ' announce connection status' property will be used, especially if the reporter thread is invoked due to a broadcaster pressing connect command (F9 in SAM, for example).
Note that this flag and property 'mismatch' is also the case with connection tone and is also a deliberate choice.