Encoder config dialog: allow changing 'announce connection status' flag. Re #122.
The 'announce connection status until connected' is meant to be configurable - by default, it is set to 'yes'. This flag should be set to 'no' (or, in encoder config set, 'yes') for archive encoders where error messages won't matter. Thus, similar to no connection tone flag, allow the new flag to be configurable in a similar fashion:
* True: announce connection status/do not stop on error.
* False: do not go pass error messages/stop when error messages are shown.