Encoder connection reporter: restore 'connecting' flag for status message announcements. Re #122.
Restore 'connecting' flag for use when announcing status messages. This flag will be set when a connection attempt is made and set to False if an error is encountered or connection is successful. This flag is needed in cases where an encoder becomes idel (or manually disconnected) while a connection attempt is being made, and is also used to prevent premature death of the reporter thread when there is a short timt pause between the connect key being pressed and encoder status showing connection messages.